first look changes
