+++ src/app/(navbar-pages)/my-borrows/page.tsx
@@ -3,12 +3,13 @@
 import { useQueryState } from "nuqs";
 import { useEffect, useState } from "react";
 
-import { BorrowRequestsTable } from "./requests/borrow-requests-table";
-import { MyBorrows } from "./requests/columns";
+import { BorrowRequestsTable, MyBorrows } from "../../../components/specific/BorrowRequestsTable";
 
 export default function MyBorrowRequests() {
-  // Available filters: "pending", "inProcess", "finished"
-  const [activeTab, setActiveTab] = useQueryState("tab", { defaultValue: "pending" });
+  const [activeTab, setActiveTab] = useQueryState<"pending" | "inProcess" | "finished">("tab", {
+    defaultValue: "pending",
+    parse: (value) => value as "pending" | "inProcess" | "finished",
+  });
   const [data, setData] = useState<MyBorrows[] | null>(null);
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState<string | null>(null);
@@ -33,21 +34,19 @@ export default function MyBorrowRequests() {
     }
   };
 
-  // Re-fetch data whenever the active tab changes.
   useEffect(() => {
     fetchData();
   }, [activeTab]);
 
-  // Pass refreshData function to child component
   return (
     <div className="mx-10 mt-6">
-      {/* Tab Navigation */}
       <div className="mb-4 flex space-x-4">
         <button
           onClick={() => setActiveTab("pending")}
           className={`rounded px-4 py-2 ${
             activeTab === "pending" ? "bg-blue-500 text-white" : "bg-gray-200 text-black"
           }`}
+          disabled={activeTab === "pending" || loading}
         >
           Angefragt
         </button>
@@ -56,6 +55,7 @@ export default function MyBorrowRequests() {
           className={`rounded px-4 py-2 ${
             activeTab === "inProcess" ? "bg-blue-500 text-white" : "bg-gray-200 text-black"
           }`}
+          disabled={activeTab === "inProcess" || loading}
         >
           Ausgeliehen
         </button>
@@ -64,6 +64,7 @@ export default function MyBorrowRequests() {
           className={`rounded px-4 py-2 ${
             activeTab === "finished" ? "bg-blue-500 text-white" : "bg-gray-200 text-black"
           }`}
+          disabled={activeTab === "finished" || loading}
         >
           Zurückgegeben
         </button>
