+++ src/components/specific/BorrowRequestsTable.tsx
@@ -1,8 +1,16 @@
-import { ColumnDef } from "@tanstack/react-table";
+"use client";
+
+import { ColumnDef, Row } from "@tanstack/react-table";
 import { de } from "date-fns/locale";
 import { Star, StarHalf, Trash, UserCheck } from "lucide-react";
 import Image from "next/image";
+import { useState } from "react";
 
+import { confirmBorrow, deleteBorrowRequest } from "@/actions/loans";
+import { ConfirmationPopup } from "@/components/general/ConfirmationPopup";
+import { RatingPopup } from "@/components/general/RatingPopup";
+import { FullDataTable } from "@/components/general/full-data-table";
+import { LoanDetailsSheet } from "@/components/specific/LoanDetailsSheet";
 import { Badge } from "@/components/ui/badge";
 import { DataTableColumnHeader } from "@/components/ui/data-table";
 import { Time } from "@/components/ui/time";
@@ -26,7 +34,6 @@ export type MyBorrows = {
   };
 };
 
-// Function to generate columns, taking openRatingPopup as a parameter
 export const getColumns = (
   openRatingPopup: (loanId: number) => void,
   handleDelete: (loanId: number, event: React.MouseEvent) => void,
@@ -37,9 +44,15 @@ export const getColumns = (
     id: "Bild",
     header: ({ column }) => <DataTableColumnHeader column={column} title="Bild" />,
     cell: ({ row }) => {
-      const pictureUrl = row.getValue("Bild") as string | null | undefined;
+      const pictureUrl = (row.getValue("Bild") as string).split(",")[0];
       return pictureUrl ? (
-        <Image src={pictureUrl} alt="Item Picture" width={50} height={50} />
+        <Image
+          src={`${process.env.NEXT_PUBLIC_MINIO_URL}/public-item-images
+/${pictureUrl}`}
+          alt="Item Picture"
+          width={50}
+          height={50}
+        />
       ) : (
         <div>No Image</div>
       );
@@ -100,16 +113,16 @@ export const getColumns = (
 
       if (isFinished) {
         status = "Zurückgegeben";
-        badgeClass = "bg-green-200 text-green-800";
+        badgeClass = "bg-green-200 hover:bg-green-200 text-green-800";
       } else if (isBorrowed) {
         status = "Ausgeliehen";
-        badgeClass = "bg-yellow-200 text-yellow-800";
+        badgeClass = "bg-yellow-200 hover:bg-yellow-200 text-yellow-800";
       } else if (isApproved) {
         status = "Angenommen";
-        badgeClass = "bg-blue-200 text-blue-800";
+        badgeClass = "bg-blue-200 hover:bg-blue-200 text-blue-800";
       } else if (isInContact) {
         status = "In Kontakt";
-        badgeClass = "bg-purple-200 text-purple-800";
+        badgeClass = "bg-purple-200 hover:bg-purple-200 text-purple-800";
       }
 
       return (
@@ -179,3 +192,99 @@ export const getColumns = (
     },
   },
 ];
+
+interface TableSectionProps {
+  data: MyBorrows[];
+  refreshData: () => void;
+}
+
+export function BorrowRequestsTable({ data, refreshData }: TableSectionProps) {
+  const [selectedLoan, setSelectedLoan] = useState<MyBorrows | null>(null);
+  const [isSheetOpen, setIsSheetOpen] = useState(false);
+  const [isRatingPopupOpen, setIsRatingPopupOpen] = useState(false);
+  const [selectedLoanId, setSelectedLoanId] = useState<number | null>(null);
+  const [isConfirmPopupOpen, setIsConfirmPopupOpen] = useState(false);
+  const [isDeletePopupOpen, setIsDeletePopupOpen] = useState(false);
+  const [loanToConfirm, setLoanToConfirm] = useState<number | null>(null);
+  const [loanToDelete, setLoanToDelete] = useState<number | null>(null);
+
+  const handleRowClick = (row: Row<MyBorrows>) => {
+    setSelectedLoan(row.original);
+    setIsSheetOpen(true);
+  };
+
+  const openRatingPopup = (loanId: number) => {
+    setSelectedLoanId(loanId);
+    setIsRatingPopupOpen(true);
+  };
+
+  const openConfirmPopup = (loanId: number, event: React.MouseEvent) => {
+    event.stopPropagation();
+    setLoanToConfirm(loanId);
+    setIsConfirmPopupOpen(true);
+  };
+
+  const openDeletePopup = (loanId: number, event: React.MouseEvent) => {
+    event.stopPropagation();
+    setLoanToDelete(loanId);
+    setIsDeletePopupOpen(true);
+  };
+
+  const handleDelete = async () => {
+    if (loanToDelete) {
+      await deleteBorrowRequest(loanToDelete);
+      refreshData();
+    }
+    setIsDeletePopupOpen(false);
+  };
+
+  const handleUserCheck = async () => {
+    if (loanToConfirm) {
+      await confirmBorrow(loanToConfirm);
+      refreshData();
+    }
+    setIsConfirmPopupOpen(false);
+  };
+
+  return (
+    <div className="mt-6">
+      <FullDataTable
+        columns={getColumns(openRatingPopup, openDeletePopup, openConfirmPopup)}
+        data={data}
+        onRowClick={handleRowClick}
+      />
+
+      {selectedLoan && (
+        <LoanDetailsSheet
+          loan={selectedLoan}
+          isOpen={isSheetOpen}
+          onClose={() => setIsSheetOpen(false)}
+        />
+      )}
+
+      <RatingPopup
+        isOpen={isRatingPopupOpen}
+        onClose={() => {
+          setIsRatingPopupOpen(false);
+          refreshData();
+        }}
+        apiUrl="/api/my-borrows/rate"
+        loanId={selectedLoanId ?? 0}
+      />
+
+      <ConfirmationPopup
+        isOpen={isConfirmPopupOpen}
+        onClose={() => setIsConfirmPopupOpen(false)}
+        message="Bist du sicher, dass du diese Ausleihe bestätigen möchtest?"
+        onConfirm={handleUserCheck}
+      />
+
+      <ConfirmationPopup
+        isOpen={isDeletePopupOpen}
+        onClose={() => setIsDeletePopupOpen(false)}
+        message="Bist du sicher, dass du diese Leihanfrage löschen möchtest?"
+        onConfirm={handleDelete}
+      />
+    </div>
+  );
+}
